Master merge internal20230905

+++ modelscope/pipelines/cv/image_panoptic_segmentation_pipeline.py
@@ -82,7 +82,7 @@ class ImagePanopticSegmentationPipeline(Pipeline):
         ids = ids[legal_indices]
         labels = np.array([id % INSTANCE_OFFSET for id in ids], dtype=np.int64)
         segms = (pan_results[None] == ids[:, None, None])
-        masks = [it.astype(np.int) for it in segms]
+        masks = [it.astype(np.int32) for it in segms]
         labels_txt = np.array(self.model.CLASSES)[labels].tolist()
         outputs = {
             OutputKeys.MASKS: masks,
